requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
# -e .
